Editing some routine enhancements (#5063)
* move routine to new file

* Move to relativedelta and add some tests

* Fix mocking

* Get match_and_return_openbb_keyword_date to 100% cov

* start testing parse script

* remove rogue prints

---------
