+++ pyproject.toml
@@ -155,6 +155,7 @@ nbmake = ">0.5.0"
 ruff = "^0.0.256"
 pytest-timeout = "^2.1.0"
 pytest-recorder = "^0.2.1"
+freezegun = "^1.2.2"
 
 [build-system]
 requires = ["setuptools<65.5.0", "poetry-core>=1.0.0"]
