+++ requirements-full.txt
@@ -91,6 +91,7 @@ fonttools==4.39.2 ; python_version >= "3.8" and python_full_version != "3.9.7" a
 formulaic==0.3.4 ; python_version >= "3.8" and python_version < "3.11" and python_full_version != "3.9.7"
 fred==3.1 ; python_version >= "3.8" and python_full_version != "3.9.7" and python_version < "3.11"
 fredapi==0.4.3 ; python_version >= "3.8" and python_full_version != "3.9.7" and python_version < "3.11"
+freezegun==1.2.2 ; python_version >= "3.8" and python_full_version != "3.9.7" and python_version < "3.11"
 frozendict==2.3.5 ; python_version >= "3.8" and python_full_version != "3.9.7" and python_version < "3.11"
 frozenlist==1.3.3 ; python_version >= "3.8" and python_full_version != "3.9.7" and python_version < "3.11"
 fsspec[http]==2023.3.0 ; python_version >= "3.8" and python_full_version != "3.9.7" and python_version < "3.11"
